增加部门注解

+++ app/src/main/java/cn/atsoft/dasheng/app/model/result/DeliveryResult.java
@@ -41,7 +41,7 @@ public class DeliveryResult implements Serializable {
     /**
      * 司机电话
      */
-    private Integer driverPhone;
+    private Long driverPhone;
     /**
      * 司机姓名
      */
