+++ app/src/main/java/cn/atsoft/dasheng/erp/model/params/OutstockApplyParam.java
@@ -35,7 +35,7 @@ public class OutstockApplyParam implements Serializable, BaseValidatingParam {
     /**
      * 司机电话
      */
-    private Integer driverPhone;
+    private Long driverPhone;
     /**
      * 司机姓名
      */
